+++ GGJ2017/Assets/Scripts/SignalWave.cs
@@ -104,7 +104,7 @@ public class SignalWave : MonoBehaviour
         var emitter = collider.GetComponent<Emitter>();
 
         var angle = Quaternion.LookRotation(Vector3.forward, _Rigidbody.velocity).eulerAngles.z;
-        emitter.SendReplicated((int) angle, transform.position, robotSpeed, InputType, Id);
+        emitter.SendReplicated(angle, transform.position, robotSpeed, InputType, Id);
 
         Destroy(gameObject);
     }
